refactor: extract font definitions and metadata generation into dedicated library modules
