+++ src/lib/fonts.ts
@@ -1,0 +1,17 @@
+import { Noto_Sans_JP, Noto_Sans_Arabic } from "next/font/google";
+
+export const notoSansJP = Noto_Sans_JP({
+  subsets: ["latin", "latin-ext"],
+  weight: ["400", "500", "700"],
+  display: "swap",
+  variable: "--font-noto-sans-jp",
+  preload: false,
+});
+
+export const notoSansArabic = Noto_Sans_Arabic({
+  subsets: ["arabic"],
+  weight: ["400", "500", "700"],
+  display: "swap",
+  variable: "--font-noto-sans-arabic",
+  preload: false,
+});
